+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Improvements
 
 - Add support for self-overlays in overlay operations (#468)
+- Improve `dissolve_within_distance` results (#494)
+- Improve performance of `join_by_location` for relation "intersects is True" (#502)
 - Add configuration option GFO_REMOVE_TEMP_FILES that can be used to avoid temp files
   being removed for debugging purposes (#480)
 - Add a context manager, TempEnv, to set temporary env variables (#481)
@@ -30,7 +32,9 @@
 
 - Set the default value of `keep_empty_geoms` to `False` for all standard operations.
   This changes the default for `make_valid` and in some cases for `simplify`. The only
-  exception is `select`, where the default stays `True` (#472, #499).
+  exception is `select`, where the default stays `True`. (#472, #499)
+- When `join_by_location` was applied, a column "spatial_relation" with the spatial
+  relation between the geometries was added. This is no longer the case. (#475)
 
 ## 0.8.1 (2024-01-13)